[seller] refactor(227): import 경로 수정
