+++ apps/seller/src/features/products/create/create-form/index.ts
@@ -3,3 +3,4 @@ export { InfoTooltip } from './info-tooltip.ui'
 export { DEFAULT_PRODUCT_OPTION } from './use-create-product-form.hook'
 export { ProductFinalPrice } from './product-final-price.ui'
 export { useCreateProductForm } from './use-create-product-form.hook'
+export { useProductCreationStore } from './product-creation.store'
